updated with section beta version

+++ app/Http/Controllers/AjaxController.php
@@ -464,9 +464,96 @@ class AjaxController extends Controller
         }    
      }
         function removeslip($refid){
+        if(Request::ajax()){
             \App\DepositSlip::where('id',$refid)->delete();
             
              return "true";
         }
+        }
         
-                                    }
+        function getstudentlist($level){
+            if(Request::ajax()){
+                
+                    $studentnames = DB::Select("select statuses.id, statuses.idno, users.lastname, "
+                        . "users.firstname, users.middlename, statuses.section  from statuses, users where statuses.idno = "
+                        . "users.idno and statuses.level = '$level' and strand = '" . Input::get("strand") ."' order by users.lastname, users.firstname, users.middlename");
+               
+                
+                $data = "";
+                $data = $data . "<table class=\"table table-stripped\"><tr><td>ID No</td><td>Name</td><td>Section</td></tr>";
+                    foreach($studentnames as $studentname){
+                        $data = $data . "<tr><td>".$studentname->idno."</td><td><span style=\"cursor:pointer\"onclick=\"setsection('" . $studentname->id . "')\">".$studentname->lastname . ", " . $studentname->firstname . " " .$studentname->middlename . "</span></td><td>" . $studentname->section . "</td></tr>"; 
+                    }
+                $data = $data."</table>";
+                
+                return $data;
+            }
+        }
+        
+        function getsection($level){
+            if(Request::ajax()){
+                $strand = Input::get("strand");
+                $sections = DB::Select("select  * from ctr_sections where level = '$level' and strand = '$strand'");
+                   $data = "";
+                   $data = $data . "<select id=\"section\" onchange=\"callsection()\" class=\"form form-control\">";
+                    foreach($sections as $section){
+                      $data = $data . "<option value= '". $section->section ."'>" .$section->section . "</option>";  
+                    }
+                   $data = $data."</select>";
+                return $data;   
+                //return "roy";
+            }
+        }
+        
+        function getsectionlist($level,$section){
+            if(Request::ajax()){
+                
+                $studentnames = DB::Select("select statuses.id, statuses.idno, users.lastname, "
+                        . "users.firstname, users.middlename, statuses.section from statuses, users where statuses.idno = "
+                        . "users.idno and statuses.level = '$level'  AND statuses.section = '$section' and strand = '" . Input::get("strand") . "' order by users.lastname, users.firstname, users.middlename");
+                $cn=1;
+                $data = "";
+                $data = $data . "<table class=\"table table-stripped\"><tr><td>ID No</td><td>CN</td><td>Name</td><td>Section</td></tr>";
+                    foreach($studentnames as $studentname){
+                        $data = $data . "<tr><td>".$studentname->idno."</td><td>" . $cn++ . "</td><td><span style=\"cursor:pointer\" onclick=\"rmsection('" . $studentname->id . "')\">".$studentname->lastname . ", " . $studentname->firstname . " " .$studentname->middlename . "</span></td><td>" . $studentname->section . "</td></tr>"; 
+                    }
+                $data = $data."</table>";
+                
+                return $data;
+                
+            }
+        }
+        
+        function setsection($id, $section){
+            if(Request::ajax()){
+            $updatesection = \App\Status::find($id);
+            $updatesection->section = $section;
+            $updatesection->update();
+            return "true";
+            }
+        }
+
+        function rmsection($id){
+            if(Request::ajax()){
+            $updatesection = \App\Status::find($id);
+            $updatesection->section = "";
+            $updatesection->update();
+            return "true";
+            }
+        }
+        
+        function getstrand($level){
+            if(Request::ajax()){
+                $strands = DB::Select("select distinct strand from ctr_payment_schedules where level = '$level'");
+                $data = "<div class=\"form form-group\"><label for=\"strand\">Select Shop/Strand</label>";
+                $data=$data. "<Select name =\"strand\" id=\"strand\" class=\"form form-control\" onchange=\"getstrandall(this.value)\" >";
+                $data=$data. "<option>--Select--</option>";
+                    foreach($strands as $strand){
+                        $data = $data . "<option value=\"". $strand->strand . "\">" . $strand->strand . "</option>";       
+                    }
+                $data = $data . "</select></div>"; 
+                return $data;
+            }
+        }
+        
+            }
